requirements.txt
streamlit
opencv-python-headless
numpy
torch
torchvision
Pillow
tqdm
ultralytics
scikit-learn
